Group network monitor stack traces
